requirements.txt
python-telegram-bot==21.4
yfinance>=0.2.38
pandas>=2.2.2
numpy>=1.26.4
ta>=0.11.0
matplotlib>=3.8.4
python-dotenv>=1.0.1
